most sequence_meta into it's own file
