+++ micromanager_gui/_gui_objects/_mda_widget.py
@@ -1,32 +1,22 @@
 from __future__ import annotations
 
 import warnings
-from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from pymmcore_plus import CMMCorePlus
 from qtpy import QtWidgets as QtW
 from qtpy import uic
 from qtpy.QtCore import QSize, Qt
 from qtpy.QtGui import QIcon
-from typing_extensions import Literal
 from useq import MDASequence
+
+from .._core import get_core_singleton
+from .._mda import SEQUENCE_META, SequenceMeta
 
 if TYPE_CHECKING:
     from pymmcore_plus.mda import PMDAEngine
 
 ICONS = Path(__file__).parent.parent / "icons"
-
-
-@dataclass
-class SequenceMeta:
-    mode: Literal["mda"] | Literal["explorer"] = ""
-    split_channels: bool = False
-    should_save: bool = False
-    file_name: str = ""
-    save_dir: str = ""
-    save_pos: bool = False
 
 
 class _MultiDUI:
@@ -90,15 +80,11 @@ class _MultiDUI:
 
 
 class MultiDWidget(QtW.QWidget, _MultiDUI):
-
-    # metadata associated with a given experiment
-    SEQUENCE_META: dict[MDASequence, SequenceMeta] = {}
-
     def __init__(self, parent=None):
         super().__init__(parent)
         self.setup_ui()
 
-        self._mmc = CMMCorePlus.instance()
+        self._mmc = get_core_singleton()
 
         self.pause_Button.released.connect(lambda: self._mmc.mda.toggle_pause())
         self.cancel_Button.released.connect(lambda: self._mmc.mda.cancel())
@@ -504,7 +490,7 @@ class MultiDWidget(QtW.QWidget, _MultiDUI):
 
         experiment = self.get_state()
 
-        self.SEQUENCE_META[experiment] = SequenceMeta(
+        SEQUENCE_META[experiment] = SequenceMeta(
             mode="mda",
             split_channels=self.checkBox_split_channels.isChecked(),
             should_save=self.save_groupBox.isChecked(),
